requirements.txt
basyx-python-sdk>=0.2.2
requests>=2.27
flask>=2.1.1
PyJWT>=2.3.0
